rearrange property decoding code
