+++ csDBPF/csDBPF/csDBPF.csproj
@@ -15,6 +15,10 @@
 	</PropertyGroup>
 
 	<ItemGroup>
+	  <None Include="..\.editorconfig" Link=".editorconfig" />
+	</ItemGroup>
+
+	<ItemGroup>
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.9.4" />
 		<PackageReference Include="MSTest.TestAdapter" Version="2.2.3" />
 		<PackageReference Include="MSTest.TestFramework" Version="2.2.3" />
